add: Role Base Permission (RBAC), using CASL. update: All Routes after using RBAC.

+++ src/api/routes/bagsRoute.js
@@ -6,6 +6,7 @@ import {
 } from '../../middlewares/validateRequest.js';
 import {
 	headersMiddleware,
+	checkPermissionMiddleware,
 	VReqToHeaderToken,
 	confirmDeleteMiddleware,
 	VReqToConfirmDelete,
@@ -33,36 +34,39 @@ bagsRoute.param('bagId', (req, res, next, bagId) =>
 	VReqToUUID(req, res, next, bagId, 'bagId')
 );
 
-//* route to get bags that user has by user => GET user muet by login
+//* route to get bags that user has by user => GET user must by login
 //* GET /user → Get all bags for a user
 //* Get bags that belong to a specific user option query
 bagsRoute.get(
 	'/',
 	VReqToHeaderToken,
 	headersMiddleware,
+	checkPermissionMiddleware('read:own', 'bag'),
 	searchMiddleware,
 	paginateMiddleware,
 	orderByMiddleware,
 	getBagsBelongsToUser
 );
 
-//* route to get bag that user has by user id => GET (params id) user muet by login
+//* route to get bag that user has by user id => GET (params id) user must by login
 //* GET /:bagId → Find a user’s bag by bag ID
 //* Get bags that belong to a specific user
 bagsRoute.get(
 	'/:bagId',
 	VReqToHeaderToken,
 	headersMiddleware,
+	checkPermissionMiddleware('read:own', 'bag'),
 	getBagBelongsToUser
 );
 
-//* route for create bag for User => POST (params id) user muet by login
+//* route for create bag for User => POST (params id) user must by login
 //* POST "/" → Create an bag for a user
 //* Create a single bag for a user
 bagsRoute.post(
 	'/',
 	VReqToHeaderToken,
 	headersMiddleware,
+	checkPermissionMiddleware('create:own', 'bag'),
 	VReqToCreateBag,
 	createBagForUser
 );
@@ -74,6 +78,7 @@ bagsRoute.put(
 	'/:bagId',
 	VReqToHeaderToken,
 	headersMiddleware,
+	checkPermissionMiddleware('update:own', 'bag'),
 	VReqToCreateBag,
 	replaceBagBelongsToUser
 );
@@ -85,30 +90,33 @@ bagsRoute.patch(
 	'/:bagId',
 	VReqToHeaderToken,
 	headersMiddleware,
+	checkPermissionMiddleware('update:own', 'bag'),
 	VReqToModifyBag,
 	modifyBagBelongsToUser
 );
 
-//* route for delete all bags By Search IF THERE that user has by user id => DELETE (params id) user muet by login
+//* route for delete all bags By Search IF THERE that user has by user id => DELETE (params id) user must by login
 //* DELETE /user/all/ → Delete all user’s bags By Search
 //* Delete all bags By Search that belong to a user
 bagsRoute.delete(
 	'/',
 	VReqToHeaderToken,
 	headersMiddleware,
+	checkPermissionMiddleware('delete:own', 'bag'),
 	searchMiddleware,
 	VReqToConfirmDelete,
 	confirmDeleteMiddleware,
 	deleteAllBagsBelongsToUser
 );
 
-//* route for delete bag user has by id of the bag => DELETE (params id) user muet by login
+//* route for delete bag user has by id of the bag => DELETE (params id) user must by login
 //* DELETE /user/:bagId → Delete a user’s bag
 //* Delete a single bag that belongs to a user
 bagsRoute.delete(
 	'/:bagId',
 	VReqToHeaderToken,
 	headersMiddleware,
+	checkPermissionMiddleware('delete:own', 'bag'),
 	deleteBagBelongsToUserById
 );
